requirements.txt
amqp==2.5.1
ansible==2.8.5
asn1crypto==0.24.0
bcrypt==3.1.7
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.1.1
Flask-HTTPAuth==3.3.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.5
MarkupSafe==1.1.1
more-itertools==7.2.0
paramiko==2.6.0
pycparser==2.19
PyNaCl==1.3.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
urllib3==1.25.6
vine==1.3.0
Werkzeug==0.16.0
zipp==0.6.0
